refactor: extract field/component builder concerns
Dedup the repeated builder setters across components:

- HasPlaceholder / HasAutoFocus / HasTabIndex / HasAutoComplete /
  HasVariant concerns in Core/Concerns replace ~17 copy-pasted methods
- Component::enumValue() replaces the inline `instanceof BackedEnum`
  coercion in method/icon/gap/align/width setters
- drop Checkbox/PasswordInput required() overrides (Field::required covers them)

No behaviour change; each component keeps the same public builder API.
