+++ src/Core/Components/Stack.php
@@ -15,17 +15,17 @@ class Stack extends ContainerComponent
 
     public function gap(Gap|string $gap): static
     {
-        return $this->prop('gap', $gap instanceof Gap ? $gap->value : $gap);
+        return $this->prop('gap', $this->enumValue($gap));
     }
 
     public function align(Align|string $align): static
     {
-        return $this->prop('align', $align instanceof Align ? $align->value : $align);
+        return $this->prop('align', $this->enumValue($align));
     }
 
     public function width(Width|string $width): static
     {
-        return $this->prop('width', $width instanceof Width ? $width->value : $width);
+        return $this->prop('width', $this->enumValue($width));
     }
 
     public function direction(string $direction): static
