+++ src/Core/Concerns/HasAutoComplete.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bambamboole\Lattice\Core\Concerns;
+
+trait HasAutoComplete
+{
+    public function autoComplete(string $autoComplete): static
+    {
+        return $this->prop('autoComplete', $autoComplete);
+    }
+}
