+++ src/Core/Concerns/HasAutoFocus.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bambamboole\Lattice\Core\Concerns;
+
+trait HasAutoFocus
+{
+    public function autoFocus(bool $autoFocus = true): static
+    {
+        return $this->prop('autoFocus', $autoFocus);
+    }
+}
