+++ src/Core/Concerns/HasPlaceholder.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bambamboole\Lattice\Core\Concerns;
+
+trait HasPlaceholder
+{
+    public function placeholder(string $placeholder): static
+    {
+        return $this->prop('placeholder', $placeholder);
+    }
+}
