+++ src/Forms/Components/NumberInput.php
@@ -2,8 +2,16 @@
 
 namespace Bambamboole\Lattice\Forms\Components;
 
+use Bambamboole\Lattice\Core\Concerns\HasAutoFocus;
+use Bambamboole\Lattice\Core\Concerns\HasPlaceholder;
+use Bambamboole\Lattice\Core\Concerns\HasTabIndex;
+
 class NumberInput extends Field
 {
+    use HasAutoFocus;
+    use HasPlaceholder;
+    use HasTabIndex;
+
     public function min(int|float $min): static
     {
         return $this->prop('min', $min);
@@ -19,24 +27,9 @@ class NumberInput extends Field
         return $this->prop('step', $step);
     }
 
-    public function placeholder(string $placeholder): static
-    {
-        return $this->prop('placeholder', $placeholder);
-    }
-
     public function slider(bool $slider = true): static
     {
         return $this->prop('slider', $slider);
-    }
-
-    public function autoFocus(bool $autoFocus = true): static
-    {
-        return $this->prop('autoFocus', $autoFocus);
-    }
-
-    public function tabIndex(int $tabIndex): static
-    {
-        return $this->prop('tabIndex', $tabIndex);
     }
 
     protected function type(): string
