+++ src/Forms/Components/PasswordInput.php
@@ -2,32 +2,17 @@
 
 namespace Bambamboole\Lattice\Forms\Components;
 
+use Bambamboole\Lattice\Core\Concerns\HasAutoComplete;
+use Bambamboole\Lattice\Core\Concerns\HasAutoFocus;
+use Bambamboole\Lattice\Core\Concerns\HasPlaceholder;
+use Bambamboole\Lattice\Core\Concerns\HasTabIndex;
+
 class PasswordInput extends Field
 {
-    public function placeholder(string $placeholder): static
-    {
-        return $this->prop('placeholder', $placeholder);
-    }
-
-    public function required(bool $required = true): static
-    {
-        return $this->prop('required', $required);
-    }
-
-    public function autoFocus(bool $autoFocus = true): static
-    {
-        return $this->prop('autoFocus', $autoFocus);
-    }
-
-    public function autoComplete(string $autoComplete): static
-    {
-        return $this->prop('autoComplete', $autoComplete);
-    }
-
-    public function tabIndex(int $tabIndex): static
-    {
-        return $this->prop('tabIndex', $tabIndex);
-    }
+    use HasAutoComplete;
+    use HasAutoFocus;
+    use HasPlaceholder;
+    use HasTabIndex;
 
     public function labelAction(string $label, string $href, ?int $tabIndex = null): static
     {
