+++ src/Forms/Components/SubmitButton.php
@@ -3,17 +3,15 @@
 namespace Bambamboole\Lattice\Forms\Components;
 
 use Bambamboole\Lattice\Core\Components\Component;
+use Bambamboole\Lattice\Core\Concerns\HasVariant;
 
 class SubmitButton extends Component
 {
+    use HasVariant;
+
     public static function make(string $label): static
     {
         return (new static)->prop('label', $label);
-    }
-
-    public function variant(string $variant): static
-    {
-        return $this->prop('variant', $variant);
     }
 
     protected function type(): string
