+++ src/Forms/Components/TextInput.php
@@ -2,33 +2,23 @@
 
 namespace Bambamboole\Lattice\Forms\Components;
 
+use Bambamboole\Lattice\Core\Concerns\HasAutoComplete;
+use Bambamboole\Lattice\Core\Concerns\HasAutoFocus;
+use Bambamboole\Lattice\Core\Concerns\HasPlaceholder;
+use Bambamboole\Lattice\Core\Concerns\HasTabIndex;
+
 class TextInput extends Field
 {
+    use HasAutoComplete;
+    use HasAutoFocus;
+    use HasPlaceholder;
+    use HasTabIndex;
+
     public function email(): static
     {
         return $this
             ->prop('type', 'email')
             ->rules(['email:rfc,filter']);
-    }
-
-    public function placeholder(string $placeholder): static
-    {
-        return $this->prop('placeholder', $placeholder);
-    }
-
-    public function autoFocus(bool $autoFocus = true): static
-    {
-        return $this->prop('autoFocus', $autoFocus);
-    }
-
-    public function autoComplete(string $autoComplete): static
-    {
-        return $this->prop('autoComplete', $autoComplete);
-    }
-
-    public function tabIndex(int $tabIndex): static
-    {
-        return $this->prop('tabIndex', $tabIndex);
     }
 
     protected function type(): string
